Implemented retrieval and display of password policy for the given user
There is not currently an API to determine which policy applies to a user. This must be determined by retrieving all policies ignoring inactive and using group membership to find the highest ranked policy applied to a group which the user is a member of.

At the moment just displaying this back to the UI.